fixed error

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -9,7 +9,7 @@
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
 	int a;
-	unsigned  int current, count = 0;
+	unsigned long int current, count = 0;
 	unsigned long int xor_v = n ^ m;
 
 	for (a = sizeof(n) * 8; a >= 0; a--)
